refactor(website): name the lib/ subsystems instead of encoding kind by position
The flat lib/ was technically correct (the convention says a loose lib/*.ts
IS an app-wide value) but the signal was positional, invisible to anyone who
has not read AGENTS.md line 323. Two subsystems had quietly crossed the point
where a folder earns its name.

lib/design/ holds the design system: recipes.ts (the class recipes and scale,
was design.ts), brand.ts, and tokens.ts (was brand-tokens.ts). The old sibling
names actively misled, nothing said design.ts and brand-tokens.ts were two
halves of one thing.

lib/ui/ holds the composed page fragments, ONE PER FILE, dissolving the single
lib/utils/ui.ts the scaffold ships. This lands the convention proposed earlier
in review: primitives in components/ui/ (owned by webjs ui add), compositions
in lib/ui/. The name is safe now because the kit mirror that used to occupy
lib/ui/ moved to lib/utils/{cn,dom}.ts, where the CLI writes it in real apps.

The growth rule is documented in AGENTS.md so this stays an extension of the
scaffold rather than a contradiction of it: start flat like the scaffold, give
a subsystem its own lib/<name>/ folder at about three related files. links.ts
and samples.ts stay loose because they have not crossed that bar.

The # alias is root-based, so moved files kept their own imports working and
only inbound references needed repointing: 14 files, plus the two reference
classes a grep for the alias misses (a test reading docs-shell by literal path
and eleven doc comments). The docs pages that teach lib/utils/ui.ts for USER
apps are deliberately untouched, the scaffold still ships that layout.

35 tests pass, webjs check clean, 12 routes verified.
