+++ website/lib/ui/docs-shell.ts
@@ -1,151 +1,10 @@
 import { html } from '@webjsdev/core';
-import { BTN_PRIMARY, BTN_GHOST, INSTALL, PANEL_CTA, SECTION, WIDE, EYEBROW } from '#lib/design.ts';
-import { DOCS_START_PATH, UI_PATH, EXAMPLE_BLOG_URL, GH_URL, DISCORD_URL, NEW_TAB } from '#lib/links.ts';
-import { brandLockup } from '#lib/brand.ts';
 
 /**
- * ui.ts: repeated MARKUP-chunk helpers (the `lib/utils/ui.ts` pattern).
- *
- * The scaffold's convention, followed here so the site matches what
- * `webjs create` generates. A design system has two kinds of helper. A
- * repeated PRIMITIVE with variants (button, badge) is a CLASS helper
- * returning a class string; those live in lib/design.ts on this site, because
- * components/ui/ is a generated mirror here rather than owned files. A
- * repeated markup CHUNK (a page header, the closing CTA, the footer, the docs
- * shell) is an `html`-fragment helper, and belongs in this file.
- *
- * These render at SSR and disappear. Nothing here registers a custom element:
- * a fragment never reaches the browser, so making one an element would ship
- * JavaScript to register a tag that does nothing.
+ * A composed page fragment (the lib/ui convention): an SSR-time function
+ * returning html. It renders and disappears. Nothing here registers a custom
+ * element; interactive elements live in components/.
  */
-
-
-/**
- * The header at the top of a long-form hub or index page.
- *
- * This existed three times, byte for byte apart from its words, on /blog,
- * /articles, and /compare. Each copy also carried an uppercase mono eyebrow
- * over the title, the device that was removed from every other page, so the
- * three pages that shared a header were also the three still carrying a
- * pattern the rest of the site had dropped. One function makes that class of
- * drift impossible: there is nowhere left for a fourth variant to appear.
- */
-export function pageHeader(title: string, lede: unknown, eyebrow?: string) {
-  return html`
-    <header class="mb-10">
-      ${eyebrow ? html`<p class=${EYEBROW}>${eyebrow}</p>` : ''}
-      <h1 class="font-serif text-hub leading-[1.05] tracking-tight text-fg mb-3">${title}</h1>
-      <p class="text-fg-muted text-sm leading-relaxed max-w-2xl">${lede}</p>
-    </header>
-  `;
-}
-
-
-/**
- * The closing call to action, composed rather than assembled per page.
- *
- * Three pages built this out of the same parts by hand, which is how the panel,
- * the install bar, and the two buttons drifted out of step with each other.
- * Passing the words in leaves nothing per-page to get wrong.
- *
- * `install` is opt-in because /why-webjs closes on the docs link alone, without
- * the command.
- */
-export function ctaPanel(opts: {
-  title: string;
-  lede: string;
-  install?: boolean;
-  primary: { href: string; label: string };
-  secondary?: { href: string; label: string; ext?: boolean };
-}) {
-  return html`
-    <section class="${SECTION} text-center" id="get-started">
-      <div class="${WIDE}">
-        <div class="max-w-3xl mx-auto p-8 md:p-12 lg:p-16 ${PANEL_CTA}">
-          <h2 class="font-display font-extrabold text-h2 leading-[1.1] tracking-[-0.03em] mt-0 mb-3">${opts.title}</h2>
-          <p class="text-fg-muted mx-auto mb-8 max-w-[52ch]">${opts.lede}</p>
-          ${opts.install === false ? '' : html`
-            <div class=${INSTALL}>
-              <span class="text-accent select-none" aria-hidden="true">$</span><copy-cmd>npm create webjs@latest my-app</copy-cmd>
-            </div>
-          `}
-          <div class="flex gap-3 justify-center flex-wrap mt-7">
-            <a class=${BTN_PRIMARY} href=${opts.primary.href}>
-              ${opts.primary.label}
-              <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2.6" stroke-linecap="round" stroke-linejoin="round" aria-hidden="true"><path d="M5 12h14"/><path d="m12 5 7 7-7 7"/></svg>
-            </a>
-            ${opts.secondary ? html`
-              <a class=${BTN_GHOST} href=${opts.secondary.href}
-                 target=${opts.secondary.ext ? '_blank' : '_self'}
-                 rel=${opts.secondary.ext ? 'noopener noreferrer' : ''}>${opts.secondary.label}${opts.secondary.ext ? NEW_TAB : ''}</a>
-            ` : ''}
-          </div>
-        </div>
-      </div>
-    </section>
-  `;
-}
-
-/**
- * The site-wide footer, rendered once by the root layout (app/layout.ts) so it
- * appears on every page, the same way the header/nav does.
- *
- * It lives here rather than inline in the layout so the chrome stays readable.
- * Pure SSR-time helper: it returns an `html` fragment and touches no client
- * globals, so importing it never ships a page to the browser.
- *
- * Anchor links point at `/#<id>` (not a bare `#<id>`) so a section anchor
- * resolves from any page, not only the home page.
- */
-export function siteFooter() {
-  return html`
-    <footer class="mt-24 border-t border-border py-16 px-6 bg-bg-subtle/30">
-      <div class="max-w-6xl mx-auto">
-        <nav class="grid grid-cols-2 md:grid-cols-5 gap-8 mb-12" aria-label="Footer">
-          <div class="flex flex-col gap-3">
-            <h4 class="text-xs font-bold uppercase tracking-wider text-fg">Product</h4>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href=${DOCS_START_PATH}>Docs</a>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href=${UI_PATH}>UI components</a>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href="/#templates">Templates</a>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href=${EXAMPLE_BLOG_URL} target="_blank" rel="noopener noreferrer">Showcase${NEW_TAB}</a>
-          </div>
-          <div class="flex flex-col gap-3">
-            <h4 class="text-xs font-bold uppercase tracking-wider text-fg">Resources</h4>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href="/what-is-webjs">What is WebJs?</a>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href="/why-webjs">Why WebJs</a>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href="/brand">Brand</a>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href="/blog">Blog</a>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href="/articles">Articles</a>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href="/changelog">Changelog</a>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href=${GH_URL + '/releases'} target="_blank" rel="noopener noreferrer">Releases${NEW_TAB}</a>
-          </div>
-          <div class="flex flex-col gap-3">
-            <h4 class="text-xs font-bold uppercase tracking-wider text-fg"><a class="no-underline text-fg hover:text-accent transition-colors" href="/compare">Compare</a></h4>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href="/compare/webjs-vs-nextjs">Next.js</a>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href="/compare/webjs-vs-lit">Lit</a>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href="/compare/webjs-vs-remix">Remix 3</a>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href="/compare/webjs-vs-astro">Astro</a>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href="/compare/webjs-vs-rails">Rails</a>
-          </div>
-          <div class="flex flex-col gap-3">
-            <h4 class="text-xs font-bold uppercase tracking-wider text-fg">Community</h4>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href=${GH_URL} target="_blank" rel="noopener noreferrer">GitHub${NEW_TAB}</a>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href=${GH_URL + '/discussions'} target="_blank" rel="noopener noreferrer">Discussions${NEW_TAB}</a>
-            <a class="text-fg-muted hover:text-accent no-underline text-sm transition-colors" href=${DISCORD_URL} target="_blank" rel="noopener noreferrer">Discord${NEW_TAB}</a>
-          </div>
-          <div class="flex flex-col gap-3">
-            <a href="/" aria-label="WebJs home" class="no-underline text-fg inline-flex w-fit transition-opacity duration-fast hover:opacity-80">${brandLockup('ftr', { height: 22 })}</a>
-            <p class="m-0 text-xs text-fg-muted leading-relaxed">The web framework for AI agents. Full-stack web components, SSR, zero build step.</p>
-          </div>
-        </nav>
-        <div class="pt-8 border-t border-border flex flex-col md:flex-row justify-between items-center gap-4 text-fg-subtle text-xs">
-          <div><a class="no-underline hover:text-accent transition-colors" href=${GH_URL + '/blob/main/LICENSE'} target="_blank" rel="noopener noreferrer">MIT License${NEW_TAB}</a></div>
-          <div class="flex items-center gap-1">Built with webjs <svg class="heart" viewBox="0 0 24 24" fill="currentColor" aria-hidden="true"><path d="M12 21.35l-1.45-1.32C5.4 15.36 2 12.28 2 8.5 2 5.42 4.42 3 7.5 3c1.74 0 3.41.81 4.5 2.09C13.09 3.81 14.76 3 16.5 3 19.58 3 22 5.42 22 8.5c0 3.78-3.4 6.86-8.55 11.54L12 21.35z"/></svg></div>
-        </div>
-      </div>
-    </footer>
-  `;
-}
 
 /**
  * The shared documentation shell: the page-tree sidebar column plus the
